# openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "Trade Platform API", "version": "0.1.0" },
  "paths": {
...
    "/api/v1/auth/signup": {
      "post": {
        "tags": ["auth"],
        "summary": "Sign Up",
        "operationId": "sign_up_api_v1_auth_signup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SignUpRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthUserResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/signin": {
      "post": {
        "tags": ["auth"],
        "summary": "Sign In",
        "operationId": "sign_in_api_v1_auth_signin_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SignInRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TokenResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthUserResponse": {
        "properties": {
          "user": { "$ref": "#/components/schemas/UserRead" },
          "token": { "$ref": "#/components/schemas/TokenResponse" }
        },
        "type": "object",
        "required": ["user", "token"],
        "title": "AuthUserResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SignInRequest": {
        "properties": {
          "email": { "type": "string", "format": "email", "title": "Email" },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Password"
          }
        },
        "type": "object",
        "required": ["email", "password"],
        "title": "SignInRequest"
      },
      "SignUpRequest": {
        "properties": {
          "email": { "type": "string", "format": "email", "title": "Email" },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "Password"
          }
        },
        "type": "object",
        "required": ["email", "password"],
        "title": "SignUpRequest"
      },
...
      "TokenResponse": {
        "properties": {
          "access_token": { "type": "string", "title": "Access Token" },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "expires_in": { "type": "integer", "title": "Expires In" }
        },
        "type": "object",
        "required": ["access_token", "expires_in"],
        "title": "TokenResponse"
      },
      "UserRead": {
        "properties": {
          "id": { "type": "integer", "title": "Id" },
          "email": { "type": "string", "format": "email", "title": "Email" },
          "is_active": { "type": "boolean", "title": "Is Active" },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": ["id", "email", "is_active", "created_at", "updated_at"],
        "title": "UserRead"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" },
          "input": { "title": "Input" },
          "ctx": { "type": "object", "title": "Context" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": { "HTTPBearer": { "type": "http", "scheme": "bearer" } }
  }
}